+++ sphinxcontrib/needs/directives/needimport.py
@@ -8,6 +8,7 @@ from docutils.parsers.rst import directives
 from docutils.parsers.rst import Directive
 from jinja2 import Template
 
+from sphinxcontrib.needs.api import add_need
 from sphinxcontrib.needs.filter_common import filter_single_need
 from sphinxcontrib.needs.utils import logger
 
@@ -26,7 +27,12 @@ class NeedimportDirective(Directive):
                    'hide': directives.flag,
                    'filter': directives.unchanged_required,
                    'id_prefix': directives.unchanged_required,
-                   'tags': directives.unchanged_required
+                   'tags': directives.unchanged_required,
+                   'style': directives.unchanged_required,
+                   'layout': directives.unchanged_required,
+                   'template': directives.unchanged_required,
+                   'pre_template': directives.unchanged_required,
+                   'post_template': directives.unchanged_required,
                    }
 
     final_argument_whitespace = True
@@ -78,19 +84,11 @@ class NeedimportDirective(Directive):
             if filter_string is None:
                 needs_list_filtered[key] = need
             else:
-                filter_context = {
-                    "tags": need["tags"],
-                    "status": need["status"],
-                    "type": need["type"],
-                    "id": need["id"],
-                    "title": need["title"],
-                    "links": need["links"],
-                    # "search": re.search,
-                    # Support both ways of addressing the description, as "description" is used in json file, but
-                    # "content" is the sphinx internal name for this kind of information
-                    "content": need["description"],
-                    "description": need["description"]
-                }
+                filter_context = {key: value for key, value in need.items()}
+
+                # Support both ways of addressing the description, as "description" is used in json file, but
+                # "content" is the sphinx internal name for this kind of information
+                filter_context['content'] = need['description']
                 try:
                     if filter_single_need(filter_context, filter_string):
                         needs_list_filtered[key] = need
@@ -107,11 +105,12 @@ class NeedimportDirective(Directive):
 
             for key, need in needs_list.items():
                 for id in needs_ids:
-                    # Manipulate links
-                    if id in need["links"]:
-                        for n, link in enumerate(need["links"]):
-                            if id == link:
-                                need["links"][n] = "".join([id_prefix, id])
+                    # Manipulate links in all link types
+                    for extra_link in env.config.needs_extra_links:
+                        if extra_link['option'] in need.keys() and id in need[extra_link['option']]:
+                            for n, link in enumerate(need[extra_link['option']]):
+                                if id == link:
+                                    need[extra_link['option']][n] = "".join([id_prefix, id])
                     # Manipulate descriptions
                     # ToDo: Use regex for better matches.
                     need["description"] = need["description"].replace(id, "".join([id_prefix, id]))
@@ -120,15 +119,49 @@ class NeedimportDirective(Directive):
         for key, need in needs_list.items():
             need["tags"] = need["tags"] + tags
 
-        template_location = os.path.join(os.path.dirname(os.path.abspath(__file__)), "needimport_template.rst")
-        with open(template_location, "r") as template_file:
-            template_content = template_file.read()
-        template = Template(template_content)
-        content = template.render(needs_list=needs_list, hide=hide, id_prefix=id_prefix)
-        self.state_machine.insert_input(content.split('\n'),
-                                        self.state_machine.document.attributes['source'])
+        # template_location = os.path.join(os.path.dirname(os.path.abspath(__file__)), "needimport_template.rst")
+        # with open(template_location, "r") as template_file:
+        #     template_content = template_file.read()
+        # template = Template(template_content)
+        # content = template.render(needs_list=needs_list, hide=hide, id_prefix=id_prefix)
+        # self.state_machine.insert_input(content.split('\n'),
+        #                                 self.state_machine.document.attributes['source'])
 
-        return []
+        need_nodes = []
+        for key, need in needs_list.items():
+            need['template'] = self.options.get("template", getattr(need, 'template', None))
+            need['pre_template'] = self.options.get("pre_template", getattr(need, 'pre_template', None))
+            need['post_template'] = self.options.get("post_template", getattr(need, 'post_template', None))
+            need['layout'] = self.options.get("layout", getattr(need, 'layout', None))
+            need['style'] = self.options.get("style", getattr(need, 'style', None))
+            need['need_type'] = need['type']
+            need['id'] = id_prefix + need['id']
+
+            need['content'] = need['description']
+            # Remove unknown options, as they may be defined in source system, but not in this sphinx project
+            extra_link_keys = [x['option'] for x in env.config.needs_extra_links]
+            extra_option_keys = list(env.config.needs_extra_options.keys())
+            default_options = ['title', 'status', 'content', 'id', 'tags', 'hide', 'template', 'pre_template',
+                               'post_template', 'collapse', 'style', 'layout', 'need_type']
+            delete_options = []
+            for option in need.keys():
+                if option not in default_options + extra_link_keys + extra_option_keys:
+                    delete_options.append(option)
+
+            for option in delete_options:
+                del (need[option])
+
+            need['docname'] = self.docname
+            need['lineno'] = self.lineno
+
+            nodes = add_need(env.app, self.state, **need)
+            need_nodes.extend(nodes)
+
+        return need_nodes
+
+    @property
+    def docname(self):
+        return self.state.document.settings.env.docname
 
 
 class VersionNotFound(BaseException):